requirements.txt
beautifulsoup4==4.12.3
blinker==1.8.2
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==42.0.8
Flask==3.0.3
Flask-Admin==1.6.1
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
greenlet==3.0.3
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
MouseInfo==0.1.3
mysql-connector==2.2.9
mysql-connector-python==9.0.0
mysql-connector-python-rf==2.2.2
pillow==10.4.0
PyAutoGUI==0.9.54
pycparser==2.22
PyGetWindow==0.0.9
PyMsgBox==1.0.9
PyMySQL==1.1.1
pyperclip==1.9.0
PyRect==0.2.0
PyScreeze==0.1.30
pytweening==1.2.0
pywhatkit==5.4
requests==2.32.3
soupsieve==2.5
SQLAlchemy==2.0.30
typing_extensions==4.12.2
urllib3==2.2.2
Werkzeug==3.0.3
wikipedia==1.4.0
WTForms==3.1.2
